refactor: add abstract class PerformanceByDeckHelperBase
