+++ models/@types/display-config.ts
@@ -1,4 +1,5 @@
-export interface PerformanceMatrixConfig {
+export interface DisplayConfig {
   onlyOurSide: boolean
   filterUser: string[]
+  onlyYourPerformance: boolean
 }
