+++ models/const/default-display-config.ts
@@ -1,0 +1,7 @@
+import { DisplayConfig } from '@/models/@types/display-config'
+
+export const defaultDisplayConfig: DisplayConfig = {
+  onlyOurSide: false,
+  filterUser: [],
+  onlyYourPerformance: false
+}
